requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0
django-allauth==65.13.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
feedparser==6.0.12
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.7.4.5
kagglehub==0.3.13
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
multidict==6.7.0
mysqlclient==2.2.7
namex==0.1.0
narwhals==2.12.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
phonenumbers==9.0.20
pillow==12.0.0
plotly==6.5.0
propcache==0.4.1
protobuf==6.33.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
xgboost==3.1.1
yarl==1.22.0
